# openapi/swagger.yaml
openapi: 3.1.0
info:
  title: HRMS Elite API v1
  version: "1.0.0"
  description: |
    OpenAPI specification for HRMS Elite API version 1.
    Includes authentication, document management and user endpoints.
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/v1/employees:
    get:
      tags: [Users]
      summary: List employees
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Array of employees
          content:
            application/json:
              examples:
                paginated:
                  $ref: '#/components/examples/PaginationExample'
    post:
      tags: [Users]
      summary: Create employee
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email]
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Employee created
  /api/v1/employees/{id}:
    get:
      tags: [Users]
      summary: Get employee by ID
      security:
        - sessionAuth: []
        - csrfToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Employee details
components:
  parameters:
    limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
      required: false
    offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
      required: false
  examples:
    PaginationExample:
      summary: Example pagination response
      value:
        data: []
        next: "http://localhost:3000/api/v1/resources?offset=10&limit=10"
        prev: null
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session authentication cookie
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF protection token
...